Add optional DHCP client process watchdog.
Watches the dhcp client process of an interface.
And calls on ifrc to check/restart the client as necessary.

Can be started anytime, or invocation added to /e/n/i file:
  post-cfg-do /etc/dhcp/client_wd -v -iwlan0

Only a single instance per interface will run.
It will re-scan for and track the client process id.
After several repeated scans, it will call on ifrc to re-'up'.

Bug 5134.