Add a day for calculated password change to prevent it being zero.
Since WB does not have functional RTC, each bootup starts at the epoch.
And the calculated day will be 0, which causes changed passwords to fail.

By adding one day, result will be that password is good for 99999-1 days.
This is a simple hack to avoid the password timeout bug.

Bug 4543.