Add support for running lighttpd via inetd.
Added [stateless] http service handling via inetd.
This will reduce system resources load for boot/idle.
However, this requires a "service-wrapper".

/
|-- etc
|   |-- inetd.conf
|   |-- init.d
|   |   |-- S50inetd
|   |   |-- S99lighttpd
|   |   `-- opt
|   |       `-- S50lighty
|   |-- lighttpd
|   |   `-- lighttpd.conf
|   `-- ssl
|       |-- certs
|       |   `-- server.crt
|       `-- server.pem
`-- usr
    `-- sbin
        |-- lighty
        `-- lighty-status

The S50inetd init-script does pre-checks on enabled services.

The wrapper-script is called 'lighty'.
It is called by inetd, or manually via shell command-line.
  (...to see usage, just run without options)

(include a '-v' in inetd.conf to see verbose messages on the console)

When a HTTP Request comes in, inetd calls lighty.
Lighty then makes a call to the lighttpd_initscript to start lighttpd.
And then parses the Headers to get the connecting address.
It then sends a HTTPS redirect to the host.

The host sees the new location and connects via HTTPS.
And user sees the dialogs for certificate and login, as necessary.

Note - The lighttpd will stay running, until 'lighty stop' is issued.

Bug 5524.