Add multipath route support for interfaces using method static.
There can only be one default route (for a gateway).
When multiple default routes are desired for multiple interfaces, they can
be added to the routing table as a 'nexthop'.

Added policy rules table support is checked firstly, and the interface is
then added to the table if not already present. (An interface's table name
is based on its ifindex value, plus an offset.)

A "route weight" value is also applied to each 'nexthop', based on the
ifindex, if not otherwise defined.

If the kernel does not include IP_ROUTE_MULTIPATH, then normal (exclusive)
default routing behavior applies.

ifrc -- v20140609

Bug 4973.