WB fw_update 2.0 -- v20131217i
Allow for local control of the update-list and transfer-list. (bug 5939)

Add '-d' option for downloading the update-list file.
Adjust handling of the update-list file.
Correct handling of the option '--url <images-server-link>'.
This option allows specifying a url/path for the image files in the list.

Updated help/usage/examples/comments.

Bug 6135