WB fw_select -- v20131217j
Rework handling of the transfer-list feature ('--transfer').

Default items to transfer are set in var TRANSFER_LIST_INCLUDE.
Additional items are specified via a transfer-list file and/or commandline.
In lieu of a specified transfer-list, the default is checked to exist:
 /tmp/alt_rootfs.transfer-list

Note, fw_update will append files from "fw.sxt" to the default file:
 /tmp/alt_rootfs.transfer-list

Items to transfer may also be passed in on the command-line with or without
a transfer-list file.  This allows quick transfers via command-line only.

The working transfer-list is scanned to prevent adding duplicates.
And a combined list is cached during copy operation as: /tmp/transfer-list~

Updated help/usage/examples/comments.

Bug 5939