Avoid using the shell group command inside a loop. Bug 6251.
This modification avoids the group command syntax bug in ash.
Changed the process id check for lighttpd, to be a function.

The shell 'busybox ash' will leak memory if using a group command
with redirection. The leak grows/persists while running in a loop.
Using function syntax avoids this shell bug.