Fix timeout/error handling and logging for various functions.
The stderr log file is shortened to fwu.err and will be appended to by
several functions.  At select stages, this file is truncated to retain
details of each image being flashed.  Content of the file is displayed
on exit_error, only if relevant to the updating process.

Also adjust/add network timeout of 4s for server pings and downloads.

Bug 7380