Fix the network-init-script background wait mechanism, for an interface.
The background wait and logging collation for each interface, is intended
for only the device/service-init related operations. And configuration is
forked so as not to hold up the init-scripts. The logged output is cached
until the interface init is completed.

This change fixes the bug for which configuration was also getting waited
on too, which was not originally intended. This problem was related to an
oddity of querying the current pid from within a subshell, always returns
the pid of the parent instead.

This "background wait" would affect behavior of "rcS_awaitscript", where
it is only intentional to wait for a specific parent init-script.

Bug 8145.