Improvements for iproute2 compliance.
Added check to ensure the ip-address and subnet are both passed in.
Note, these values can be of the old config style or in CIDR notation.
Additionally, the network, wildcard, broadcast and netmask bits are each
calculated in succession, as necessary.

The legacy check/set-value for broadcast, network, subnet ..are removed.

Corrected the ip-addr-modify for inet4 and the ip rule/route cases.

Bug 7064