Add patch for handling roamed events.
The busybox-1.23.x ifplugd had some changes that prevented the old patch
from being applied cleanly. One of the changes is for variable declaration
that seems dubious, but interferes with new code, so prior code retained.
Another change is elimination of a (minimal) delay_time recycle ..for which
may or may not be good to retain. So will try keeping this change for now.

Combined/adjusted the 1.21.x patch for handling netlink state w/iface error
handling, since these are related and need to be together anyway. And this
patch is now renamed as *-networking-ifplugd-nl-state.

Bug 8276