Fix udhcpc.sh setting of log-level for udhcpc debug and capture.
Moved the eval of DHCP_PARAMS to occur before udhcpc_start() is called.
This allows the log file setting to be in effect for capturing.

Adjusted log-level for verboseness vs. for debugging udhcpc operation.

The verboseness level is intended for the wrapper and udhcpc.script.
A level of 1 is the default, and passing in a '-v' sets level 2.
More '-v's passed in allow to set verbosity (3 max) of udhcpc.

Bug 8399.