laird: build: Remove /run in post_build step to avoid conflicts
Remove /run so we can avoid a conflict with our method of redirecting /run
to /tmp via a symlink. If a package creates a subdirectory here, we break
the build when doing this because the act of copying our rootfs additions
conflict when it can't change the directory to a symlink.

Bug: 8932