laird: wbXXn: merge wb45/wb50 files that are not different into common
It's recognized that this will pose a problem for the WB40 as it
continues to stagnate and the needs for the WB45 and WB50 might not
match it as it is dependent on earlier kernel and system software. This
is not seen as a problem as the current buildroot and kernel aren't
adjusted to deal with the WB40, and as such the WB40 targets aren't
expected to build. Management has had us halt further upgrades of the 40
system software, so it's not deemed as an issue.

	renamed:    wb50n/rootfs-additions/etc/dhcp/udhcpd.wlan0.conf -> rootfs-additions-common/etc/dhcp/udhcpd.wlan0.conf
	renamed:    wb50n/rootfs-additions/etc/fw_env.config -> rootfs-additions-common/etc/fw_env.config
	renamed:    wb50n/rootfs-additions/etc/init.d/opt/S91g_ether -> rootfs-additions-common/etc/init.d/opt/S91g_ether
	renamed:    wb50n/rootfs-additions/etc/modprobe.d/ath6k.conf -> rootfs-additions-common/etc/modprobe.d/ath6k.conf
	renamed:    wb50n/rootfs-additions/etc/network/interfaces -> rootfs-additions-common/etc/network/interfaces
	renamed:    wb50n/rootfs-additions/etc/summit/.empty -> rootfs-additions-common/etc/summit/.empty
	deleted:    wb45n/rootfs-additions/etc/dhcp/udhcpd.wlan0.conf
	deleted:    wb45n/rootfs-additions/etc/fw_env.config
	modified:   wb45n/rootfs-additions/etc/hostapd/hostapd.conf
	deleted:    wb45n/rootfs-additions/etc/init.d/opt/S91g_ether
	deleted:    wb45n/rootfs-additions/etc/modprobe.d/ath6k.conf
	deleted:    wb45n/rootfs-additions/etc/network/interfaces
	deleted:    wb45n/rootfs-additions/etc/summit/.empty
	deleted:    wb50n/rootfs-additions/etc/init.d/opt/S50sshd

bug 8698