laird: package/pkg-kconfig: define KCONFIG_EDITORS_MAKE_ENV
Originally from Tal Zilcer from the buildroot mainling list

This fixes the ability to do a linux-xconfig build target.

NOTE: this wasn't the accepted solution, but there is upstreamed fixes for
this problem, 1a91c82f8a8e2de1ab95794429dfa90b78d51670. This commit caused
conflicts and required a large set to resolve properly, the commit picked
here is cleaner for now, but needs to backed out properly next time we
upgrade buildroot past 2015.05

Bug: 9140

--

When building a configuration editor we assume its for the HOST.
However some packages might define MAKE_ENV as TARGET_MAKE_NEV.
Using this variable adds to beginning of $(PATH),
the directory $(HOST_DIR)/usr/bin.

This cause us to not use pkg-config from our host machine but the one
from $(HOST_DIR)/usr/bin. This pkg-config does not support QT.
Now when we try to run for example linux-xconfig we fail since QT is not found.

To overcome this we defined a dedicated variable for the configuration editors

Signed-off-by: Tal Zilcer <talz@ezchip.com>