Laird: bridge.sh: Fix detection of wireless interface as a bridge_port.
During bridge port checking, latch the $phy80211 flag as true when a
wireless interface is given.  And then when installing the L2 rules,
check if this flag is set true or assume it is false.

This fix allows the bridge_ports to be specified in any order.  So that
when any of bridge_ports are wireless, then the additional L2 rules will
be installed to support wireless bridging.

Bug 9320