package/ca-certificates: depend on host-python or host-python3
The ca-certificates build process runs a Python script that needs at
least Python 2.7. While Buildroot requires Python as a hard dependency
on the build system, we don't require Python >= 2.7.

So in order to ensure that a Python >= 2.7 is installed, this commit
makes the ca-certificates package depend either on host-python or
host-python3.

Fixes:

 http://autobuild.buildroot.net/results/a2a4fed293c836b9cf63ff2aaa463b0704dec07e/
 http://autobuild.buildroot.net/results/bbed4afa8e30382b8892062f31ba64cbb0ea14e4/
 http://autobuild.buildroot.net/results/1af5562be3c0d233cea81834a898f7ac6ae48271/

Cc: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Martin Bark <martin@barkynet.com>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
[Thomas: rewrite commit log.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>