laird: wireless.sh: Allow usage of alternate ath6kl_usb module.
The driver and module path may be set near the top of the script, to use
appropriate hardware device module connected via USB.  The iface checking
routine has been modified so can work with *_sdio or *_usb modules during
start.  And for when stopping, the correct modules will be removed.

bug 9187