laird: nodejs: Allow selection as a host tool
Adding nodejs as a host tool allows packages which use npm to be
installed for the target such as angular.
If the host tool is selected the latest supported version is defined.

A host tool wrapper is added to allow packages to install with npm
easily with $(HOST_NPM)

Bug 10391