laird: msd45n: change msd45n to use old sdc compiler
The newer compiler sets the EABI softfloat bit in the elf header.
This seems to cause issues on EABIhf platforms, but code compiled with
the old compiler doesn't set this bit, yet still runs. By changing the
compiler back to the old one for just the MSD45 build, we hope to fix
compatibility with other platforms.

Note that this causes a hybrid model on the wb45n - our closed source
binaries are built with the old compiler and flag, but the rest of the
wb code is built with the newer one. This is necessitated by newer code
like DCAL not building with the old compiler. Testing shows that this all
seems to work.

Bug: 10388