laird: Include sourcemaps and aot compile in web_lcm_gatwick
Keeping source maps for the time being as its useful for debugging
in the browser. We turn AOT on to reduce page generation on the fly.

We are now able to call the local installation of ng instead of having
a host version to keep in sync with.

Production also zips up a copy of the files we can safely remove the .gz

Bug 10611