laird: rdvk: Add support to rdvk for testing LWB's FCC and ETSI only modes
This adds the ability to disable CRDA's regulatory database signature
checking.  This allows the rdvk to use CRDA with custom regulatory
databases we generate without having any key infrastructure in place. Two
custom databases, FCC and ETSI only, are added to the rdvk to be used with
the LWB to mimic the built in FCC and ETSI only LWB backport drivers.

Bug: 10454