laird: wb50n_rdvk: get firmware from laird_firmware package
We were keeping the firmware in each board/rootfs-additions directory and
manually managing the firmware was getting difficult as we would have to
update it in multiple spots. By creating a firmware repository and a new
buildroot package, we minimize the maintenance issue, and gain significant
flexibility in configurations. It also lets other types of builds use the
one repository without having to pull a full buildroot tree.

Now that we have this, remove all the firmware from the wb50n-rdvk and add
the relevant configuration to handle this. Note that the wb50n-rdvk needs
all firmware we support so we have selected all of the configuration flags.

Bug: 10686