laird: som60: generate a combined UBI with genimage
We will roll the UBI with genimage from the components.
The image contains the squashfs ro rootfs, the verity hash of the block,
and the ubifs rootfs overlay.
No need to create the rootfs overlay with an empty set of data as this
is done on the live system.
Genimage handles the ubinizing so we can remove calling it directly as
well.

The partition sizes are an estimate of the size needed, 80MB RO, 4MB
verity, remaining RW expanded on first boot.
As it stands for a 85MB ro verity grows to ~760K so the 1MB will contain
it all.

Bug 12760