laird: som60: add boot script that includes the verity hash and salt
As part of our secure booting we need to link our verified root to the
kernel, to do this we generate a bootscript for uboot which includes
the hash and salt used by the verity rootfs.
This script constructs the bootparams based on the rootfs info.
It can be run with the example command
'nand read 0x21000000 0x000E0000 0x00800000; source 0x21000000:script;
bootm'
The standard bootcmd will continue to be able to load an un verified
system for the time being.

Bug 12760