laird: Add sbom-gen target in buildroot Makefile
sbom-gen target depends on legal-info target and calls includes csv-to-txt
and legal-info-to-sbom script. These scripts fixes the warning generated
by the legal-info target of the package sourced from externals dir.Creates
a SBOM which can be ingested by CVE-checker. The generated SBOM is outputted
in buildroot/output/(target)/images.

bug 12793