laird: Correct the naming of the .swu file for the SOM60
The .swu file isn't being pushed to Devops at the completion of the build
because the common make script looks for a file name with the date string.
Removing the version string from the file name as the binaries collected
will be in a versioned directory on the share.

Bug 13725