laird: som60: Add support for flashing som60 nand images from som60sd
In order to update the nand from an sd image, swupdate needs the ubi device
configured with the correct ubi volumes.  The prep_nand_for_update script
has to be run before swupdate is run.  It cannot be run as a preinstall
script in swupdate because swupdate scans for the ubi volumes before it
runs any preinstall scripts.

A uboot environment binary is necessary as the uboot environment needs
to be initialized properly.  This normally happens with a 'saveenv' in
the uboot shell.  Instead, from userspace, we will flash a uboot env binary
generated with the 'mkenvimage' command.

Bug 13297