package/libiberty: new package
From: Adam Duskett <Aduskett@gmail.com>

Some packages, like prelink-cross, will want to use libiberty but do
not bundle their own instance (which is good!).

However, libiberty is made for being bundled in packages: all GNU
packages that use libiberty (gcc, binutils, gdb et al...) all have
their own bundled variant. This means that there is no official
upstream for libiberty; the closest being as part of the
combined binutils-gdb tree.

So we introduce a new host-only package, that installs just libiberty
from a binutils released tarball.

Again, as libiberty is usually bundled in packages, it usually only
install a static version. Furthermore, it does not obey to the usual
--enable-shared and --disable-static flags; it only ever builds a
static version.

Furthermore, this library is not built with -fPIC, but some packages
may pick it to build shared objects. This is the case for host-gdb,
for example, which accidentally picks that library instead of its
internal one.

So, rather than fix the various gdb versions and variants we can use,
we ensure that the libiberty we install is usable in shared objects,
and we ensure it is always built before host-gdb.

Bug 15879