Add node-ssl-shim package
In the past, Red Hat released a set of patches to Node.js to support integration
of older (i.e., 1.0.2) and FIPS versions of OpenSSL. With newer versions of Node.js
(>= 12.18.4), these patches have been collected into the 'node-ssl-shim' project
(https://github.com/sclorg/node-ssl-shim) which combines the patches into a
library which can be linked against.

This commit pulls in the 'node-ssl-shim' package in order to provide support for
newer versions of Node.js.

Bug 19956