CVE-2020-23922: Add bounds checking on indices in DumpScreen2RGB()
CVE-2020-23922 is described as "an issue was discovered in giflib through 5.1.4.
DumpScreen2RGB in gif2rgb.c has a heap-based buffer over-read."

The issue is on line 294 - the value passed into the function is used to index
into the array (which is also passed into the function) without checking that it
is not out of bounds. The risk here is low as the DumpScreen2RGB() function
looks to be only used internally, but nevertheless, the patch in this commit
adds an extra check to make sure the ScreenWidth and ScreenHeight are valid
values.

At present, this issue has not yet been addressed in the giflib sources (see the
relevant bug here: https://sourceforge.net/p/giflib/bugs/151/); therefore, this
patch is needed.

Bug 20353