buildroot: node-ssl-shim - Fix install staging commands
The install staging step of the build process for the node-ssl-shim package
was failing on CI by pulling in the wrong float ABI (soft instead of hard).
The change here pulls switches to calling 'install' in the package
Makefile instead of using the 'make install' command of the source Makefile.

Bug 20428