docs/manual: Remove link to emdebian and replace with debian port page
In the list of full distribution, the FAQ contains a link to Emdebian.
This project stopped receiving updates in 2014 and the main web page does not exist anymore.
This replace the entry with a link to the Debian ports page.

Signed-off-by: Yannick Brosseau <yannick.brosseau@gmail.com>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
(cherry picked from commit db2b8a1ce22e9ee458532d2900d4c8cbe96d9ba0)
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>