buildroot: ig60llsd: Add dependencies to support running GGv2
This change adds the OpenJDK package at version 11 - the version currently
recommended by AWS for GGv2:
https://docs.aws.amazon.com/greengrass/v2/developerguide/setting-up.html#greengrass-v2-requirements

The OpenJDK packages depends on Xorg 7, so that package is included as well.

As found with IG60 GG build, the BusyBox 'ps' applet does not support the -x
option which is used by GGv2 (see Bug 20772). Therefore, the BusyBox 'ps'
implementation is replaced with 'procps-ng' which includes a number of process
related applications. To accomodate this, a new, ig60-specific BusyBox config
file is added to the Buildroot sources which disables the BusyBox applets whose
functionality is now provided by 'procps-ng'.

Additionally, GGv2 requires the use of the 'sudo' package, so this is included
as well.

Bug 21392