swupdate: Systemd Integration
This change enables better integration of swupdate running as a systemd
service.

Firstly, the swupdate-progress service is preset to be disabled as we are not
using this functionality and monitor update progress ourselves (e.g., in
igupd, etc.).

Secondly, an environment variable configuration file has been added to set the
run-time configuration file to use (/etc/swupdate.cfg) and the accepted select
software image sets for use with swupdate.

Lastly, since systemd is now responsible for starting the swupdate binary, the
fw_update script has been updated to call the 'swupdate-client' application
instead of 'swupdate' directly. Additionally, as the 'swupdate-client'
application does not support downloading an update file from a remote URL, a
call to curl (or wget if curl is not available) was added to download the file
and stream it to 'swupdate-client'.

Also, weblcm-python previously installed a systemd service unit file for
swupdate that conflicts with the one provided by swupdate itself. This change
removes that service unit file, replicates the proper generation of the
SWUPDATE_ARGS environment variable to support signed images, and adds a
dependency on systemd to weblcm-python.

Bug 21723