buildroot: weblcm-python: Enable support for stunnel control in WebLCM
This change adds a new config option to weblcm-python to enable the new
'stunnel' module (BR2_PACKAGE_WEBLCM_ENABLE_STUNNEL_CONTROL).

Also, the stunnel package has been updated to install the systemd service file.

Additionally, stunnel was added as a backup target in the makefile for
lrd-encrypted-storage-toolkit and an override was added for stunnel to allow
it to inherit the keyring.

Bug 22594